pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.23",
    "connect-markdown-renderer>=3.0.0",
    "langchain-text-splitters>=0.3.2",
    "langchain>=0.3.10",
    "langchain-community>=0.3.10",
    "langchain-core>=0.3.22",
    "langchain-google-genai>=2.0.6",
    "langchain-groq>=0.2.1",
    "langgraph>=0.2.56",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
    "transformers>=4.46.3",
    "tiktoken>=0.8.0",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "prompt-toolkit>=3.0.48",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
